Add MiniDrawer component and update navigation
