+++ src/components/BarNavigation.tsx
@@ -1,37 +1,20 @@
 import * as React from "react";
-import { styled, useTheme, Theme, CSSObject } from "@mui/material/styles";
+import { styled, Theme, CSSObject } from "@mui/material/styles";
 import MuiDrawer from "@mui/material/Drawer";
 import MuiAppBar, { AppBarProps as MuiAppBarProps } from "@mui/material/AppBar";
-import Toolbar from "@mui/material/Toolbar";
-import Typography from "@mui/material/Typography";
-import Divider from "@mui/material/Divider";
-import IconButton from "@mui/material/IconButton";
-import MenuIcon from "@mui/icons-material/Menu";
-import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-import ChevronRightIcon from "@mui/icons-material/ChevronRight";
-import AccountCircleIcon from "@mui/icons-material/AccountCircle";
-import Badge from "@mui/material/Badge";
-import NotificationsActiveIcon from "@mui/icons-material/NotificationsActive";
-import Popover from "@mui/material/Popover";
-import List from "@mui/material/List";
-import {
-	Grid,
-	ListItemButton,
-	ListItemText,
-	Menu,
-	MenuItem,
-	Stack,
-} from "@mui/material";
-import Tooltip from "@mui/material/Tooltip";
-import { useSectionHeaders } from "./SectionHeader";
-import SectionHeader from "./SectionHeader";
 import { createBrowserHistory } from "history";
+import MiniDrawer from "./MiniDrawer";
+import Drivers from "../pages/Drivers";
+import Passengers from "../pages/Passengers";
+import Settings from "../pages/Settings";
+import Reports from "../pages/Reports";
+import Trips from "../pages/Trips";
 
 // Define the width of the drawer
 const drawerWidth = 240;
 
 // Create a browser history object
-const history = createBrowserHistory();
+export const history = createBrowserHistory();
 
 // CSS mixins for the opened and closed states of the drawer
 const openedMixin = (theme: Theme): CSSObject => ({
@@ -56,7 +39,7 @@ const closedMixin = (theme: Theme): CSSObject => ({
 });
 
 // Styled component for the header of the drawer
-const DrawerHeader = styled("div")(({ theme }: { theme: Theme }) => ({
+export const DrawerHeader = styled("div")(({ theme }: { theme: Theme }) => ({
 	display: "flex",
 	alignItems: "center",
 	justifyContent: "flex-end",
@@ -69,7 +52,7 @@ interface AppBarProps extends MuiAppBarProps {
 	open?: boolean;
 }
 
-const AppBar = styled(MuiAppBar, {
+export const AppBar = styled(MuiAppBar, {
 	shouldForwardProp: (prop) => prop !== "open",
 })<AppBarProps>(({ theme, open }) => ({
 	zIndex: theme.zIndex.drawer + 1,
@@ -88,7 +71,7 @@ const AppBar = styled(MuiAppBar, {
 }));
 
 // Custom Drawer component that extends MuiDrawer
-const Drawer = styled(MuiDrawer, {
+export const Drawer = styled(MuiDrawer, {
 	shouldForwardProp: (prop: string) => prop !== "open",
 })(({ theme, open }: { theme: Theme; open: boolean }) => ({
 	width: drawerWidth,
@@ -105,191 +88,17 @@ const Drawer = styled(MuiDrawer, {
 	}),
 }));
 
-// Props for the MiniDrawer component
-type MiniDrawerProps = {
-	children?: React.ReactNode;
+const BarNavigation: React.FC = () => {
+	return (
+		<MiniDrawer>
+			{/* <Login /> */}
+			{/* <Drivers /> */}
+			{/* <Passengers /> */}
+			{/* <Settings /> */}
+			{/* <Reports /> */}
+			{/* <Trips /> */}
+		</MiniDrawer>
+	);
 };
 
-// Main component
-export default function MiniDrawer({ children }: MiniDrawerProps) {
-	const theme = useTheme();
-	const [open, setOpen] = React.useState(false);
-	const {
-		sectionHeaders,
-		openAllSections,
-		closeAllSections,
-		toggleSectionOpen,
-	} = useSectionHeaders();
-
-	// Event handler for opening the drawer
-	const handleDrawerOpen = () => {
-		setOpen(true);
-		openAllSections();
-	};
-
-	// Event handler for closing the drawer
-	const handleDrawerClose = () => {
-		setOpen(false);
-		closeAllSections();
-	};
-
-	// State and event handlers for the notification popover
-	const [anchorEl, setAnchorEl] = React.useState<HTMLElement | null>(null);
-	const handleMenu = (event: React.MouseEvent<HTMLElement>) => {
-		setAnchorEl(event.currentTarget);
-	};
-	const handleClose = () => {
-		setAnchorEl(null);
-	};
-
-	// State and event handlers for the notification popover
-	const [notificationAnchorEl, setNotificationAnchorEl] =
-		React.useState<HTMLElement | null>(null);
-	const handleNotificationClick = (event: React.MouseEvent<HTMLElement>) => {
-		setNotificationAnchorEl(event.currentTarget);
-	};
-	const handleNotificationClose = () => {
-		setNotificationAnchorEl(null);
-	};
-	const openNotification = Boolean(notificationAnchorEl);
-
-	// Event handler for clicking the app bar heading
-	const handleAppBarHeadingClick = () => {
-		history.push(`/`);
-	};
-
-	// Function to calculate the notification count
-	const notificationCount = () => {
-		return 1 + 1;
-	};
-
-	return (
-		<Stack>
-			<AppBar
-				position="fixed"
-				open={open}
-			>
-				<Toolbar>
-					<IconButton
-						color="inherit"
-						aria-label="open drawer"
-						onClick={handleDrawerOpen}
-						edge="start"
-						sx={{ marginRight: 5, ...(open && { display: "none" }) }}
-					>
-						<MenuIcon />
-					</IconButton>
-					<Typography
-						variant="h6"
-						noWrap
-						component="div"
-						onClick={handleAppBarHeadingClick}
-					>
-						ROMODO
-					</Typography>
-					<div style={{ marginLeft: "auto" }}>
-						<IconButton
-							size="large"
-							color="inherit"
-							aria-label="notifications"
-							onClick={handleNotificationClick}
-						>
-							<Badge
-								badgeContent={notificationCount()}
-								color="warning"
-							>
-								<NotificationsActiveIcon />
-							</Badge>
-						</IconButton>
-						<Popover
-							open={openNotification}
-							anchorEl={notificationAnchorEl}
-							onClose={handleNotificationClose}
-							anchorOrigin={{ vertical: "bottom", horizontal: "left" }}
-							transformOrigin={{ vertical: "top", horizontal: "left" }}
-						>
-							<List>
-								<ListItemButton>
-									<ListItemText primary="Notification 1" />
-								</ListItemButton>
-								<ListItemButton>
-									<ListItemText primary="Notification 2" />
-								</ListItemButton>
-								<ListItemButton>
-									<ListItemText primary="Notification 3" />
-								</ListItemButton>
-							</List>
-						</Popover>
-						<IconButton
-							size="large"
-							aria-label="account"
-							aria-controls="menu-appbar"
-							aria-haspopup="true"
-							onClick={handleMenu}
-							color="inherit"
-						>
-							<AccountCircleIcon />
-						</IconButton>
-						<Menu
-							id="menu-appbar"
-							anchorEl={anchorEl}
-							anchorOrigin={{ vertical: "top", horizontal: "right" }}
-							keepMounted
-							transformOrigin={{ vertical: "top", horizontal: "right" }}
-							open={Boolean(anchorEl)}
-							onClose={handleClose}
-						>
-							<MenuItem onClick={handleClose}>Manage Profile</MenuItem>
-							<MenuItem onClick={handleClose}>Logout</MenuItem>
-						</Menu>
-					</div>
-				</Toolbar>
-			</AppBar>
-			<Grid container>
-				<Grid item>
-					<Drawer
-						theme={theme}
-						variant="permanent"
-						open={open}
-					>
-						<DrawerHeader>
-							<IconButton onClick={handleDrawerClose}>
-								{theme.direction === "rtl" ? (
-									<ChevronRightIcon />
-								) : (
-									<ChevronLeftIcon />
-								)}
-							</IconButton>
-						</DrawerHeader>
-						<Divider />
-						<List>
-							{sectionHeaders.map(
-								({ text, icon, nestedItems, open }, index) => (
-									<Tooltip
-										title={text}
-										key={index}
-									>
-										<SectionHeader
-											key={index}
-											text={text}
-											icon={icon}
-											nestedItems={nestedItems || []}
-											open={open}
-											handleClick={() => {
-												if (nestedItems.length > 0) {
-													setOpen(true);
-													toggleSectionOpen(text);
-												}
-											}}
-										/>
-									</Tooltip>
-								)
-							)}
-						</List>
-					</Drawer>
-				</Grid>
-				<Grid item>{children}</Grid>
-			</Grid>
-		</Stack>
-	);
-}
+export default BarNavigation;
